requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==2.5.1
Flask-Migrate==3.1.0
Flask-Login==0.6.2
scikit-learn==1.1.3
pandas==1.5.3
numpy==1.24.0
lightgbm==3.3.2
tensorflow==2.10.1
Keras==2.10.0
opencv-python==4.6.0.66
matplotlib==3.6.2
Pillow==9.2.0
torch==1.13.1
cupy-cuda116==11.5.0
python-dotenv==0.21.0
